+++ app/src/main/java/eu/kanade/tachiyomi/util/chapter/ChapterUtil.kt
@@ -153,7 +153,7 @@ class ChapterUtil {
             return chapters.size > 20
         }
 
-        const val scanlatorSeparator = " & "
+        const val scanlatorSeparator = " [.] "
 
         fun getScanlators(scanlators: String?): List<String> {
             if (scanlators.isNullOrBlank()) return emptyList()